Correcting i32 type

+++ mpi/mpi_cpu_nrm2.cpp
@@ -21,7 +21,7 @@ do_setup(const benchmark::State & state)
     auto n = state.range(0);
     x = new double[n];
 
-    for (i32 i = 0; i < n; ++i) {
+    for (i64 i = 0; i < n; ++i) {
         x[i] = std::rand();
     }
 }
@@ -43,7 +43,7 @@ mpi_cpu_nrm2_weak(benchmark::State & state)
         comm.barrier();
         auto t0 = mpi::wall_time();
         double nrm2 = 0.;
-        for (i32 i = 0; i < n; ++i) {
+        for (i64 i = 0; i < n; ++i) {
             nrm2 += x[i] * x[i];
         }
         double gnrm2;
